change to double for timer

+++ include/CUDA_Bench/gemm/gemm_cutlass_launch_int.cuh
@@ -71,7 +71,7 @@ int gemm_cutlass_launch_int()
     gpuErrchk(cudaDeviceSynchronize());
     cudaEventRecord(time_stop, 0);
     cudaEventSynchronize(time_stop);
-    float elapsedTime;
+    double elapsedTime;
     cudaEventElapsedTime(&elapsedTime, time_start, time_stop);
     cudaEventDestroy(time_start);
     cudaEventDestroy(time_stop);
